requirements.txt
pillow
pandas
numpy
matplotlib
torch
